Refactor auth code structure
