+++ src/auth/auth.controller.ts
@@ -10,8 +10,8 @@ import {
 import { AuthService } from './auth.service';
 import { RegisterDto } from './dto/register.dto';
 import { User } from '../users/entities/user.entity';
-import { LocalAuthGuard } from './local-auth.guard';
-import { SessionAuthGuard } from './session-auth.guard';
+import { LocalAuthGuard } from './guards/local-auth.guard';
+import { SessionAuthGuard } from './guards/session-auth.guard';
 import { Request } from 'express';
 import {
   ApiBadRequestResponse,
