+++ src/auth/decorators/user.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { User } from '../users/entities/user.entity';
+import { User } from '../../users/entities/user.entity';
 
 export const ReqUser = createParamDecorator(
   (data: unknown, ctx: ExecutionContext): User | null => {
