+++ src/auth/guards/roles.guard.ts
@@ -6,8 +6,8 @@ import {
   UnauthorizedException,
 } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
-import { Role } from '../users/entities/role.enum';
-import { ROLES_KEY } from './roles.decorator';
+import { Role } from '../../users/entities/role.enum';
+import { ROLES_KEY } from '../decorators/roles.decorator';
 
 @Injectable()
 export class RolesGuard implements CanActivate {
