+++ src/orders/payment-methods/payment-methods.controller.ts
@@ -12,7 +12,7 @@ import { PaymentMethodsService } from './payment-methods.service';
 import { PaymentMethod } from '../entities/payment-method.entity';
 import { PaymentMethodDto } from '../dto/payment-method.dto';
 import { Role } from '../../users/entities/role.enum';
-import { Roles } from '../../auth/roles.decorator';
+import { Roles } from '../../auth/decorators/roles.decorator';
 import {
   ApiBadRequestResponse,
   ApiCreatedResponse,
