+++ src/products/attribute-types/attribute-types.controller.ts
@@ -10,7 +10,7 @@ import {
 } from '@nestjs/common';
 import { AttributeType } from '../entities/attribute-type.entity';
 import { AttributeTypesService } from './attribute-types.service';
-import { Roles } from '../../auth/roles.decorator';
+import { Roles } from '../../auth/decorators/roles.decorator';
 import { Role } from '../../users/entities/role.enum';
 import { AttributeTypeDto } from '../dto/attribute-type.dto';
 import {
