+++ src/products/products.controller.ts
@@ -20,7 +20,7 @@ import {
 import { ProductsService } from './products.service';
 import { Product } from './entities/product.entity';
 import { Role } from '../users/entities/role.enum';
-import { Roles } from '../auth/roles.decorator';
+import { Roles } from '../auth/decorators/roles.decorator';
 import { ProductCreateDto } from './dto/product-create.dto';
 import { ProductUpdateDto } from './dto/product-update.dto';
 import { AttributeDto } from './dto/attribute.dto';
