+++ src/wishlists/wishlists.controller.ts
@@ -8,11 +8,11 @@ import {
   Post,
 } from '@nestjs/common';
 import { WishlistsService } from './wishlists.service';
-import { ReqUser } from '../auth/user.decorator';
+import { ReqUser } from '../auth/decorators/user.decorator';
 import { User } from '../users/entities/user.entity';
 import { Wishlist } from './entities/wishlist.entity';
 import { Role } from '../users/entities/role.enum';
-import { Roles } from '../auth/roles.decorator';
+import { Roles } from '../auth/decorators/roles.decorator';
 import { WishlistCreateDto } from './dto/wishlist-create.dto';
 import { WishlistUpdateDto } from './dto/wishlist-update.dto';
 import {
